a solid model with lines showing the faces (working)
